--- Procfile
@@ -1,2 +1,0 @@
-web: npm start
-worker: node worker.js
