+++ web.js
@@ -31,13 +31,13 @@ app.use(function (req, res, next) {
     next(err);
 });
 
-// error handlers
-app.use(function (err, req, res, next) {
-    res.status(err.status || 500);
-    res.json({
-        message: err.message,
-        error: {}
-    });
-});
+//// error handlers
+//app.use(function (err, req, res, next) {
+//    res.status(err.status || 500);
+//    res.json({
+//        message: err.message,
+//        error: {}
+//    });
+//});
 
 module.exports = app;
